test util with `TempFixture`
