+++ src/test_utils.rs
@@ -1,0 +1,37 @@
+use std::path::{Path, PathBuf};
+
+/// A struct that contains a temporary directory and a path to a file in that directory.
+pub struct TempFixture {
+    _temp_dir: tempfile::TempDir,
+    temp_path: PathBuf,
+}
+
+impl TempFixture {
+    /// Creates a copy of the test file in a temporary directory.
+    /// Returns the struct `TempFixture` that contains the temp dir (for clean-up on `drop`) as well as the path to the file.
+    ///
+    /// This can potentially be removed when <https://github.com/OSGeo/gdal/issues/6253> is resolved.
+    pub fn fixture(name: &str) -> Self {
+        let path = std::path::Path::new("fixtures").join(name);
+
+        let _temp_dir = tempfile::tempdir().unwrap();
+        let temp_path = _temp_dir.path().join(path.file_name().unwrap());
+
+        std::fs::copy(&path, &temp_path).unwrap();
+
+        Self {
+            _temp_dir,
+            temp_path,
+        }
+    }
+
+    pub fn path(&self) -> &Path {
+        &self.temp_path
+    }
+}
+
+impl AsRef<Path> for TempFixture {
+    fn as_ref(&self) -> &Path {
+        self.path()
+    }
+}
